refactor: App into 5 files
